upgrading drawable

+++ adam/src/main/java/com/schematical/adam/drawable/opengl/AdamOpenGLDrawable.java
@@ -4,31 +4,31 @@ package com.schematical.adam.drawable.opengl;
  * Created by user1a on 10/7/13.
  */
 public class AdamOpenGLDrawable {
-    public int x = 0;
-    public int y = 0;
-    public int z = 0;
+    public float x = 0;
+    public float y = 0;
+    public float z = 0;
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public int getZ() {
+    public float getZ() {
         return z;
     }
 
-    public void setZ(int z) {
+    public void setZ(float z) {
         this.z = z;
     }
 }
